==> 02_Creating_new_variables.ipynb <==
# Creating new variables

import numpy as np
import pandas as pd
from pandas import DataFrame, Series

kobe = pd.DataFrame.from_csv("./data/kobe_searching_for_new_variables.csv", sep=",")

## Creating variables crossing 'action_type' with shot zone and distance variables ('shot_zone_area', 'shot_zone_basic', 'shot_zone_range', and 'shot_distance')

Analyzing the data on Tableau we have observed that these are the only variables that exhibit remarkable variations from the average shot accuracy. 
It is possible to find the charts of the accuracy percentage according to the crossing of these categories on Tableau book 'kobe_searching_new_variables.twb' in the directory 'visualization'.

kobe["at_sza"] = kobe["action_type"] + "-" + kobe["shot_zone_area"]

kobe["at_sza"].unique()

len(kobe["at_sza"].unique())

kobe["at_szb"] = kobe["action_type"] + "-" + kobe["shot_zone_basic"]

kobe["at_szb"].unique()

len(kobe["at_szb"].unique())

kobe["at_szr"] = kobe["action_type"] + "-" + kobe["shot_zone_range"]

kobe["at_szr"].unique()

len(kobe["at_szr"].unique())

kobe["at_sd"] = kobe["action_type"] + "-" + kobe["shot_distance"].astype(str)

kobe["at_sd"].unique()

len(kobe["at_sd"].unique())

514 values of categorical variable out of a sample of about 30,000 events might be too many. We shall remember it when using these variables in the prediction model.

## Creating trend variable: percentage of made shots over the previous three months

#### Note: incomplete.

After observing data on Tableau, I'm not sure a trend variable can be relevant for a prediction model. Anyway let's try to create it as an exercise.

We drop the submission set.

kobe_trend = kobe[pd.notnull(kobe["shot_made_flag"])]

We create a new column adding up 'game_year' and 'game_month' labels.

kobe_trend["game_year_month"] = (kobe_trend["game_year"].astype(str)) + "-" + (kobe_trend["game_month"].astype(str))

Now we create a dataframe after doing a groupby to find out the number of shots for every year/month.

kobe_trend_totalshots = pd.DataFrame({"total_count" : kobe_trend.groupby(["game_year_month"]).size()}).reset_index()

kobe_trend_totalshots.head(3)

We repeat the process to calculate the number of shots in for every year/month.

kobe_trend_madeshots = pd.DataFrame({"made_count" : kobe_trend.groupby(["game_year_month", "shot_made_flag"]).size()}).reset_index()

kobe_trend_madeshots = kobe_trend_madeshots[kobe_trend_madeshots["shot_made_flag"]==1]

kobe_trend_madeshots.head(3)

Now we merge the dataframes.

kobe_trend_merged = pd.merge(kobe_trend_madeshots, kobe_trend_totalshots, on="game_year_month")

kobe_trend_merged.head(3)

We turn the values of 'game_year_month' into datetime values in order to work with time series.

kobe_trend_merged.dtypes

kobe_trend_merged["game_year_month"] = pd.to_datetime(kobe_trend_merged["game_year_month"])

We sort the values in chronological order.

kobe_trend_merged = kobe_trend_merged.sort("game_year_month")

We standarize the day for all dates.

kobe_trend_merged["game_year_month"] = kobe_trend_merged["game_year_month"].apply(lambda dt: dt.replace(day=1))

len(kobe_trend_merged)

print kobe_trend_merged["game_year_month"]

At this point I have achieved to create a function that returns a list of lists with the three previous dates if they appear in 'game_year_month' of kobe_trend_merged db.

def get_previous3months(df):
    p3m_ls = df["game_year_month"].apply(lambda date: [date - pd.DateOffset(months=1), date - pd.DateOffset(months=2), date - pd.DateOffset(months=3)])
    return p3m_ls

previous3months_lists = get_previous3months(kobe_trend_merged)

previous3months_lists

By manually selecting the row of 'kobe_trend_merged' I extract the chart with the values of the previous three months (if there is any).

previous3months_table = kobe_trend_merged[kobe_trend_merged["game_year_month"].isin(previous3months_lists[24])]

previous3months_table

Next I calculate the accuracy rate (the sought value).

previous3months_table[["made_count", "total_count"]] = previous3months_table[["made_count", "total_count"]].astype(float)
accuracy_prev3m = 1000 * np.around(sum(previous3months_table["made_count"]) / sum(previous3months_table["total_count"]), 3)
accuracy_prev3m

I imagine I should add a for comprehension function to extract in a loop the data contained in the table and create an if-else condition to asign 0.0 value to the dates of 'kobe_trend_merged['game_year_month']' with no sample of shots during the previous three months. From here it should be easy to calculate the accuracy rate.

_______________________________________________________________________________________________________________________

We export the dataset with the new variables to use it in the next step.

kobe.to_csv("./data/kobe_new_variables.csv", sep= ",")
