==> 03_Preparing_data_for_prediction_model.ipynb <==
# Preparing data for Logistic Regression

import numpy as np
import pandas as pd
from pandas import DataFrame, Series

kobe = pd.DataFrame.from_csv("./data/kobe_new_variables.csv", sep=",")

## Converting variables to category type for better summarization

About the benefits of turning variables into categorical: http://pandas.pydata.org/pandas-docs/stable/categorical.html

kobe.dtypes

Before turning variables into categories, we drop 'season' (string values composed by numbers) out. If we turn it into a category, as we should, we are not sure how the LogisticRegression will react (it's neither int nor float). Probably we should make dummy variables with it, but it is not worth, as we already have 'game_year' variable (int).

kobe = kobe.drop("season", axis=1)

First of all we turn non numerical variables that seem to fit into 'category' dtype.

kobe["action_type"] = kobe["action_type"].astype("category")
kobe["combined_shot_type"] = kobe["combined_shot_type"].astype("category")
kobe["playoffs"] = kobe["playoffs"].astype("category")
kobe["shot_made_flag"] = kobe["shot_made_flag"].astype("category")
kobe["shot_zone_area"] = kobe["shot_zone_area"].astype("category")               
kobe["shot_zone_basic"] = kobe["shot_zone_basic"].astype("category")             
kobe["shot_zone_range"] = kobe["shot_zone_range"].astype("category")
kobe["opponent"] = kobe["opponent"].astype("category")
kobe["local"] = kobe["local"].astype("category")
kobe["at_sza"] = kobe["at_sza"].astype("category")                          
kobe["at_szb"] = kobe["at_szb"].astype("category")    
kobe["at_szr"] = kobe["at_szr"].astype("category")    
kobe["at_sd"] = kobe["at_sd"].astype("category")                           

We check the categories of 'shot_made_flag' target.

kobe["shot_made_flag"].cat.categories

We turn also some numerical variables into categories, specifying an order on them.

kobe["game_event_id"] = pd.Categorical(kobe["game_event_id"], ordered = True)
kobe["game_id"] = pd.Categorical(kobe["game_id"], ordered = True)
kobe["period"] = pd.Categorical(kobe["period"], ordered = True)
kobe["game_year"] = pd.Categorical(kobe["game_year"], ordered = True)
kobe["game_month"] = pd.Categorical(kobe["game_month"], ordered = True)
kobe["game_dayofweek"] = pd.Categorical(kobe["game_dayofweek"], ordered = True)
kobe["game_dayofyear"] = pd.Categorical(kobe["game_dayofyear"], ordered = True)

Let's take a look at some basic statistics by dtype.

kobe.describe(include=["category"])

kobe.describe(include=["number"])

## Making dummy variables

LogisticRegression model of sklearn  doesn't accept string values. OneHotEncoder is the sklearn standart approach to convert categorial features into numerical (http://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.OneHotEncoder.html), but it was returning errors reading str values. LabelEncoder is much easier to implement, but it might not be effective when applying the created values to the prediction model (http://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.LabelEncoder.html). So let's use pandas get_dummies function to create dummy variables.

old_cat_variables = ["action_type", "combined_shot_type", "shot_zone_area", "shot_zone_basic", "shot_zone_range", "opponent"]

kobe_old_cat_vars = kobe[old_cat_variables]

kobe_old_cat_vars.shape

kobe_sparse_old_catvars = pd.get_dummies(kobe_old_cat_vars)

kobe_sparse_old_catvars.shape

new_cat_variables = ["at_sza", "at_szb", "at_szr", "at_sd"]

kobe_new_cat_vars = kobe[new_cat_variables]

kobe_new_cat_vars.shape

kobe_sparse_new_catvars = pd.get_dummies(kobe_new_cat_vars)

kobe_sparse_new_catvars.shape

## Reducing the number of viariables with Boruta and Random Forest Classifier

#### Note: incomplete.

Currently we're handling slightly fewer than 1,000 variables. There is the risk that our model will not detect well the significant variables. So we'll apply an all-relevant feature selection method - Boruta (feature reduction done slavishly following the example you can find here: https://github.com/danielhomola/boruta_py

from sklearn.ensemble import RandomForestClassifier
from boruta_py import BorutaPy

kobe_target = kobe["shot_made_flag"]

kobe_target

kobe_socv = pd.concat([kobe_sparse_old_catvars, kobe_target], axis=1)

kobe_socv["shot_made_flag"]

kobe_socv = kobe_socv.dropna()

kobe_socv.shape

y=kobe_socv.pop("shot_made_flag")
X=kobe_socv.copy()

First we will instantiate an estimator that Boruta will use. Then we will instantiate a Boruta Object.

rf = RandomForestClassifier(n_jobs=-1, class_weight="auto", max_depth=8)
feat_selector = BorutaPy(rf, n_estimators="auto", verbose=2)

Once built, we can use this object to identify the relevant features in our dataset.

feat_selector.fit(X,y)

We check the selected features

print feat_selector.support_

We check ranking of features

print feat_selector.ranking_

We call transform() on X to filter it down to selected features

X_filtered = feat_selector.transform(X)

_________________________________________________________________________________________________________________________

Since we were not able to reduce the number of variables, it's highly possible that the model will be less effective. So we'll try with different aggregations of the variables that we are currently handling. So we export three different dataset: the original with variables containing numerical data, the second with sparsed distribution of the original categorical variables. and the third with sparsed distribution of the newly created categorical variables.

kobe = kobe[["game_event_id", "game_id", "loc_x", "loc_y", "period", "playoffs", "shot_distance","shot_made_flag", \
     "seconds_to_period_end", "accurate_shot_distance", "game_year", "game_month", "game_dayofweek", "game_dayofyear", \
      "game_dayofweek", "game_dayofyear", "local"]]

kobe.to_csv("./data/kobe_num_variables.csv", sep= ",")

kobe_sparse_old_catvars.to_csv("./data/kobe_old_cat_variables.csv", sep= ",")

kobe_sparse_new_catvars.to_csv("./data/kobe_new_cat_variables.csv", sep= ",")
