==> 04_Prediction_Model.ipynb <==
# Prediction model: Logistic Regression

import numpy as np
import pandas as pd
from pandas import DataFrame, Series

kobe_num = pd.DataFrame.from_csv("./data/kobe_num_variables.csv", sep=",")

kobe_num.shape

kobe_num.head(3)

kobe_old_cat = pd.DataFrame.from_csv("./data/kobe_old_cat_variables.csv", sep=",")

kobe_old_cat.shape

kobe_old_cat.head(3)

kobe_new_cat = pd.DataFrame.from_csv("./data/kobe_new_cat_variables.csv", sep=",")

kobe_new_cat.shape

kobe_new_cat.head(3)

## Splitting the submission data

kobe_num_submission = kobe_num[pd.isnull(kobe_num["shot_made_flag"])]

We add 'shot_made_flag' to 'kobe_old_cat' and 'kobe_new_cat'

kobe_target = kobe_num["shot_made_flag"]

kobe_old_cat = pd.concat([kobe_old_cat, kobe_target], axis=1)

kobe_new_cat = pd.concat([kobe_new_cat, kobe_target], axis=1)

And we create the two remaining submission data

kobe_old_cat_submission = kobe_old_cat[pd.isnull(kobe_old_cat["shot_made_flag"])]

kobe_new_cat_submission = kobe_new_cat[pd.isnull(kobe_new_cat["shot_made_flag"])]

print kobe_num_submission.shape
print kobe_old_cat_submission.shape
print kobe_new_cat_submission.shape

We split the dataset from submission data

kobe_num = kobe_num[pd.notnull(kobe_num["shot_made_flag"])]
kobe_old_cat = kobe_old_cat[pd.notnull(kobe_old_cat["shot_made_flag"])]
kobe_new_cat = kobe_new_cat[pd.notnull(kobe_new_cat["shot_made_flag"])]

print len(kobe_num)
print len(kobe_old_cat)
print len(kobe_new_cat)

## Logistic Regression

Finally we finally run logistic regression on the three data set ('kobe_num', 'kobe_old_cat', 'kobe_new_cat'). Let see if we can find an aggregation that can be accurate.

from sklearn.linear_model import LogisticRegression
from sklearn.cross_validation import train_test_split
from sklearn import metrics
from sklearn.cross_validation import cross_val_score

#### Let's start with the first data, that should be not particuarly accurate.

We need to flatten `y` into a 1-D array, so that scikit-learn will properly understand it as the response variable.

y = np.ravel(kobe_num["shot_made_flag"])

kobe_num = kobe_num.drop("shot_made_flag", axis=1)

X = kobe_num

model = LogisticRegression()
model = model.fit(X, y)

model.score(X, y)

55.4% accuracy doesn't particularly seem good.

y.mean()

44.6% of the shots were in. We could obtain 68% accuracy by always predicting "no".

Let examine the coefficients to see what we learn.

pd.DataFrame(zip(X.columns, np.transpose(model.coef_)))

## Model Evaluation using a validation set

So far we have trained and tested on the same set. Now we split the data into a training set and a testing set to evaluate the model.

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)
model2 = LogisticRegression()
model2.fit(X_train, y_train)

Now we predict the class labels for the test set.

predicted = model2.predict(X_test)
print predicted

We will also generate the class probabilities.

probs = model2.predict_proba(X_test)
print probs

The classifier is predicting a 1 (shot made) any time the probability in the second column is greater than 0.5.

Now we generate some evaluation metrics.

print metrics.accuracy_score(y_test, predicted)
print metrics.roc_auc_score(y_test, probs[:, 1])

The accuracy is 55.8%, which is similar to what we experienced when training and predicting on the same data.

We can also see the confusion matrix and a classification report with other metrics.

print metrics.confusion_matrix(y_test, predicted)
print metrics.classification_report(y_test, predicted)

## Model Evaluation Using Cross-validation

Now we evaluate the model using 10-fold cross-validation, to see if the accuracy holds up more rigorously.

scores = cross_val_score(LogisticRegression(), X, y, scoring='accuracy', cv=10)
print scores
print scores.mean()

At least it's still performing at around 55.4 accuracy.

#### Let's try with the sparse matrix of the original and theoretically more relevant variables.

y = np.ravel(kobe_old_cat["shot_made_flag"])

kobe_old_cat = kobe_old_cat.drop("shot_made_flag", axis =1)

X = kobe_old_cat

model = LogisticRegression()
model = model.fit(X, y)
model.score(X, y)

68.2% accuracy. It seems that these variables are more useful.

pd.DataFrame(zip(X.columns, np.transpose(model.coef_)))

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)
model2 = LogisticRegression()
model2.fit(X_train, y_train)

predicted = model2.predict(X_test)
print predicted

probs = model2.predict_proba(X_test)
print probs

print metrics.accuracy_score(y_test, predicted)
print metrics.roc_auc_score(y_test, probs[:, 1])

Similar accuracy to what we experienced when training and predicting on the same data.

print metrics.confusion_matrix(y_test, predicted)
print metrics.classification_report(y_test, predicted)

At least this model it's able to predict shot in with some accuracy

scores = cross_val_score(LogisticRegression(), X, y, scoring='accuracy', cv=10)
print scores
print scores.mean()

Fine. It's still performing at around 68% accuracy.

#### Now we'll try with the sparse matrix of the variables newly created crossing types of actions with different segmentation of position on court. There are 871 variables: without a feature selection it's highly probable the model will be not accurate

y = np.ravel(kobe_new_cat["shot_made_flag"])

kobe_new_cat = kobe_new_cat.drop("shot_made_flag", axis=1)

X = kobe_new_cat

model = LogisticRegression()
model = model.fit(X, y)
model.score(X, y)

pd.DataFrame(zip(X.columns, np.transpose(model.coef_)))

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)
model2 = LogisticRegression()
model2.fit(X_train, y_train)

predicted = model2.predict(X_test)
print predicted

probs = model2.predict_proba(X_test)
print probs

Something less accurate then the second model.

print metrics.accuracy_score(y_test, predicted)
print metrics.roc_auc_score(y_test, probs[:, 1])

print metrics.confusion_matrix(y_test, predicted)
print metrics.classification_report(y_test, predicted)

scores = cross_val_score(LogisticRegression(), X, y, scoring='accuracy', cv=10)
print scores
print scores.mean()

#### The second model was a little bit better than the third. Let's try to merge all the variables formerly present in the dataset to see if the model manages to be tighter.

kobe_num = pd.DataFrame.from_csv("./data/kobe_num_variables.csv", sep=",")

kobe_old_cat = pd.DataFrame.from_csv("./data/kobe_old_cat_variables.csv", sep=",")

kobe_num = kobe_num[pd.notnull(kobe_num["shot_made_flag"])]

kobe = pd.concat([kobe_num, kobe_old_cat], axis=1)

kobe_submission = kobe[pd.isnull(kobe["shot_made_flag"])]

kobe = kobe[pd.notnull(kobe["shot_made_flag"])]

y = np.ravel(kobe["shot_made_flag"])

kobe = kobe.drop("shot_made_flag", axis=1)

X = kobe

model = LogisticRegression()
model = model.fit(X, y)
model.score(X, y)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)
model2 = LogisticRegression()
model2.fit(X_train, y_train)

predicted = model2.predict(X_test)
print predicted

probs = model2.predict_proba(X_test)
print probs

print metrics.accuracy_score(y_test, predicted)
print metrics.roc_auc_score(y_test, probs[:, 1])

print metrics.confusion_matrix(y_test, predicted)
print metrics.classification_report(y_test, predicted)

scores = cross_val_score(LogisticRegression(), X, y, scoring='accuracy', cv=10)
print scores
print scores.mean()

Definitely not a good model. Now we save the submission dataset corresponding to the second dataset to recreate the model in the following script and predict the probability of made shot on submission dataset.

kobe_old_cat_submission.to_csv("./data/kobe_model_submission.csv", sep= ",")
